+++ yaybu/core/runner.py
@@ -19,6 +19,7 @@ import logging
 import logging.handlers
 import collections
 import traceback
+import subprocess
 
 import yay
 
@@ -26,59 +27,13 @@ from yaybu.core.shell import Shell
 from yaybu.core import change
 from yaybu.core import resource
 from yaybu.core import error
+from yaybu.core import remote
 
 logger = logging.getLogger("runner")
 
 class LoaderError(Exception):
     pass
 
-class RunContext:
-
-    simulate = False
-    ypath = ()
-    verbose = 0
-    html = None
-
-    def __init__(self, opts=None):
-        self.path = []
-        self.ypath = []
-        if opts is not None:
-            logger.debug("Invoked with ypath: %r" % opts.ypath)
-            logger.debug("Environment YAYBUPATH: %r" % os.environ.get("YAYBUPATH", ""))
-            self.simulate = opts.simulate
-            self.ypath = opts.ypath
-            self.verbose = opts.verbose
-            if opts.html is not None:
-                self.html = open(opts.html, "w")
-        if "PATH" in os.environ:
-            for term in os.environ["PATH"].split(":"):
-                self.path.append(term)
-        if "YAYBUPATH" in os.environ:
-            for term in os.environ["YAYBUPATH"].split(":"):
-                self.ypath.append(term)
-
-    def locate(self, paths, filename):
-        if filename.startswith("/"):
-            return filename
-        for prefix in paths:
-            candidate = os.path.realpath(os.path.join(prefix, filename))
-            logger.debug("Testing for existence of %r" % (candidate,))
-            if os.path.exists(candidate):
-                return candidate
-            logger.debug("%r does not exist" % candidate)
-        raise ValueError("Cannot locate file %r" % filename)
-
-    def locate_file(self, filename):
-        """ Locates a file by referring to the defined yaybu path. If the
-        filename starts with a / then it is absolutely rooted in the
-        filesystem and will be returned unmolested. """
-        return self.locate(self.ypath, filename)
-
-    def locate_bin(self, filename):
-        """ Locates a binary by referring to the defined yaybu path and PATH. If the
-        filename starts with a / then it is absolutely rooted in the
-        filesystem and will be returned unmolested. """
-        return self.locate(self.ypath + self.path, filename)
 
 
 class Runner(object):
@@ -122,21 +77,7 @@ class Runner(object):
             simhandler.setFormatter(simformatter)
             simlog.addHandler(simhandler)
 
-    def run(self):
-        parser = optparse.OptionParser()
-        parser.add_option("-s", "--simulate", default=False, action="store_true")
-        parser.add_option("-p", "--ypath", default=[], action="append")
-        parser.add_option("", "--log-facility", default="2", help="the syslog local facility number to which to write the audit trail")
-        parser.add_option("", "--log-level", default="info", help="the minimum log level to write to the audit trail")
-        parser.add_option("-d", "--debug", default=False, help="switch all logging to maximum, and write out to the console")
-        parser.add_option("-l", "--logfile", default=None, help="The filename to write the audit log to, instead of syslog. Note: the standard console log will still be written to the console.")
-        parser.add_option("-v", "--verbose", default=1, action="count", help="Write additional informational messages to the console log. repeat for even more verbosity.")
-        parser.add_option("-H", "--html", default=None, help="Instead of writing progress information to the console, write an html progress log to this file.")
-
-        opts, args = parser.parse_args()
-        if len(args) != 1:
-            parser.print_help()
-            return 1
+    def run(self, opts, args):
         try:
             if opts.debug:
                 opts.html = False
@@ -158,5 +99,27 @@ class Runner(object):
             return e.returncode
 
 def main():
-    return Runner().run()
+    parser = optparse.OptionParser()
+    parser.add_option("-s", "--simulate", default=False, action="store_true")
+    parser.add_option("-p", "--ypath", default=[], action="append")
+    parser.add_option("", "--log-facility", default="2", help="the syslog local facility number to which to write the audit trail")
+    parser.add_option("", "--log-level", default="info", help="the minimum log level to write to the audit trail")
+    parser.add_option("-d", "--debug", default=False, help="switch all logging to maximum, and write out to the console")
+    parser.add_option("-l", "--logfile", default=None, help="The filename to write the audit log to, instead of syslog. Note: the standard console log will still be written to the console.")
+    parser.add_option("-v", "--verbose", default=1, action="count", help="Write additional informational messages to the console log. repeat for even more verbosity.")
+    parser.add_option("-H", "--html", default=None, help="Instead of writing progress information to the console, write an html progress log to this file.")
+    parser.add_option("-h", "--host", default=None, action="store", help="A host to remotely run yaybu on")
+    parser.add_option("--remote", default=False, action="store_true", help="Run yaybu.protocol client on stdio")
+
+    opts, args = parser.parse_args()
+    if len(args) != 1:
+        parser.print_help()
+        return 1
+
+    if opts.host:
+        r = remote.RemoteRunner()
+    else:
+        r = Runner()
+
+    return r.run()
